pitch-visualizer/requirements.txt
# ── Core web framework ────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
jinja2>=3.1.3
pydantic>=2.7.0
httpx>=0.27.0
python-dotenv>=1.0.1


# ── Image handling ────────────────────────────────────────────────────────────
Pillow>=10.3.0

# ── NLP / segmentation ────────────────────────────────────────────────────────
nltk>=3.8.1

# ── Google Gemini for Prompt Enhancement ─────────────────────────────────────
google-generativeai>=0.5.0

